renamed header and cpp file
